refactor: data.js 폴더 위치 변경 - utils 폴더로 옮겨주었습니다
